OS X build fixes

+++ Sources/GL/Platform/OSX/input_device_provider_osxkeyboard.h
@@ -81,11 +81,11 @@ public:
 
 	/// \brief Returns the x position of the device.
 	/** <p>Only valid for mouse.</p>*/
-	int get_x() const { return 0; }
+	float get_x() const { return 0; }
 
 	/// \brief Returns the y position of the device.
 	/** <p>Only valid for mouse.</p>*/
-	int get_y() const { return 0; }
+	float get_y() const { return 0; }
 
 	/// \brief Returns true if a tablet stylus is in proximity (hovering close enough over the tablet surface).
 	bool in_proximity() const { return false; }
